Code/requirements.txt
fastapi
uvicorn
sqlalchemy
sendgrid
python-jose
passlib[bcrypt]
python-multipart
pydantic
